添加信息提示条，中奖烟花效果

+++ app/level.vue
@@ -1,4 +1,5 @@
 <script>
+	import key from 'keymaster';
 	import DB from './db';
 	import WinManager from './win';
 	export default {
@@ -19,12 +20,25 @@
 			this.wins = WinManager.all();
 			// 初始化当前抽奖奖项
 			this.prize = this.prizes[0];
+
+			this.prizes.forEach((p, i) => {
+				console.log(p, i);
+				// bind events
+				key(i+1+'', () => {
+					this.prize = this.prizes[i];
+					// returning false stops the event and prevents default browser events
+					return false;
+				});
+			})
+
+
 		},
 
 		methods: {
 
 			// 删除获奖者
 			deleteWinner(name) {
+				if(!confirm('此操作不能恢复，确定删除吗？')){ return; }
 				WinManager.delete(name);
 				this.winOfCurrPrize = WinManager.allNameOfPrize(this.prize);
 			},
@@ -45,18 +59,18 @@
 
 <template>
 	<div class="award-level" v-if="prize">
-		<h2>
+		<h2 class="clearfix">
 			<span class="prize"
 				v-for="item,index in prizes"
 				:class="{ active: item.name === prize.name }"
 				@click="selectPrize(index)">{{item.name}}</span>
+			<code class="fr">{{winOfCurrPrize.length}}/{{prize.num}}</code>
 		</h2>
 		<div class="award-winner">
-			<span v-for="name in winOfCurrPrize">
+			<span class="animated fadeInDown" v-for="name,index in winOfCurrPrize">
 				{{name}}
 				<a href="javascript:;" class="delete" @click="deleteWinner(name)">x</a>
 			</span>
-			<p>{{winOfCurrPrize.length}}/{{prize.num}}</p>
 		</div>
 	</div>
 </template>
@@ -64,6 +78,16 @@
 <style lang='stylus' scoped>
 	.award-level
 		color #fff;
+		code
+			display: inline-block
+			background: url('./img/list.png') 3px 0 no-repeat
+			height: 25px
+			line-height: 21px
+			color: #fff
+			border: 2px solid #cecece
+			border-radius: 3px
+			padding: 1px 5px 1px 25px
+			margin-top: 5px
 	.prize
 		margin-right: 20px
 		color: #888
@@ -74,12 +98,13 @@
 	.award-winner
 			min-height: 78px
 			> span
-				padding: 10px 15px
+				font-size: 3rem;
+				padding: 12px 15px 0px 15px
 				margin: 10px 15px 10px 0
-				background: #df0007
 				color: #fff
 				font-weight: 700
 				display: inline-block
+				background: url(./img/crown.png) 2px -10px no-repeat #df0007
 				position: relative
 			.delete
 				position: absolute
@@ -88,5 +113,4 @@
 				text-decoration: none
 				font-size: 0.8rem
 				color: #fff
-
 </style>
